% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{import.kilo4}
\alias{import.kilo4}
\title{Import raw kilosort4 and trial data}
\usage{
import.kilo4(
  folder_path,
  stim_data_file,
  recording_Fq_MATLAB = 30303,
  trial_time_start = -100,
  trial_time_end = 400,
  verbose = TRUE
)
}
\arguments{
\item{folder_path}{Path to the kilosort4 output files}

\item{stim_data_file}{Path to the stimulus event data file}

\item{recording_Fq_MATLAB}{Sampling frequency of the MATLAB recording}

\item{trial_time_start}{Time to begin trial (ms), relative to stimulus onset}

\item{trial_time_end}{Time to end trial (ms), relative to stimulus onset}

\item{verbose}{Print report on imported data?}
}
\value{
A list with two data frames: spikes (rows are spikes) and stim (rows are stimulus onset times)
}
\description{
Base function to import raw kilosort4 and trial data.
}
