% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess.kilo4}
\alias{preprocess.kilo4}
\title{Preprocess kilosort4 data}
\usage{
preprocess.kilo4(
  trial_time_start = -100,
  trial_time_end = 2020,
  recording.folder = "data",
  pure_trials_only = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{trial_time_start}{Time to begin trial (ms), relative to stimulus onset}

\item{trial_time_end}{Time to end trial (ms), relative to stimulus onset}

\item{recording.folder}{List of paths to the kilosort4 output folders}

\item{pure_trials_only}{Keep only trials with no overlap?}

\item{verbose}{Print report on imported data?}
}
\value{
A list with three elements: spikes (data frame with one row per spike), timeXtrial (list of matrices with rows as sample times and columns as trials), and cluster.key (data frame with one row per neuron)
}
\description{
Function to preprocess kilosort4 data for use with neurons package functions.
}
